fix: close phase Y sprint Y1 QA findings
